+++ app/src/main/kotlin/me/xizzhu/android/joshua/App.kt
@@ -20,6 +20,8 @@ import android.app.Activity
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.HasActivityInjector
+import me.xizzhu.android.joshua.core.analytics.Analytics
+import me.xizzhu.android.joshua.core.analytics.android.FirebaseAnalyticsProvider
 import me.xizzhu.android.joshua.core.logger.Log
 import me.xizzhu.android.joshua.core.logger.android.CrashlyticsLogger
 import javax.inject.Inject
@@ -39,6 +41,7 @@ class App : BaseApp(), HasActivityInjector {
         appComponent.inject(this)
 
         Log.addLogger(CrashlyticsLogger())
+        Analytics.addProvider(FirebaseAnalyticsProvider(this))
     }
 
     override fun activityInjector(): AndroidInjector<Activity> = dispatchingAndroidInjector