+++ app/src/test/kotlin/me/xizzhu/android/joshua/core/repository/TranslationRepositoryTest.kt
@@ -178,6 +178,9 @@ class TranslationRepositoryTest : BaseUnitTest() {
     @Test
     fun testDownloadTranslation() {
         runBlocking {
+            doReturn(0L).`when`(translationRepository).elapsedRealtime()
+            doReturn(null).`when`(translationRepository).buildParams(any(), anyLong(), anyLong(), anyLong())
+
             val channel = Channel<Int>()
             `when`(remoteTranslationService.fetchTranslation(channel, RemoteTranslationInfo.fromTranslationInfo(MockContents.kjvTranslationInfo)))
                     .thenReturn(RemoteTranslation(RemoteTranslationInfo.fromTranslationInfo(MockContents.kjvTranslationInfo),